added src folder and moved App.js to src folder
