+++ src/App.js
@@ -22,7 +22,7 @@ export default class App extends Component<Props> {
     return (
       <View style={styles.container}>
         <Text style={styles.welcome}>Welcome to React Native!</Text>
-        <Text style={styles.instructions}>To get started, edit App.js</Text>
+        <Text style={styles.instructions}>App.js is now moved under 'src' folder.</Text>
         <Text style={styles.instructions}>{instructions}</Text>
       </View>
     );
